+++ lib/extended_email_reply_parser/parsers/i18n_de.rb
@@ -1,0 +1,12 @@
+module ExtendedEmailReplyParser
+  class Parsers::I18nDe < Parsers::Base
+
+    def parse
+      except_in_visible_block_quotes do
+        hide_everything_after ["Von: ", "Gesendet: ", "An: "]
+        hide_everything_after ["Am ", "schrieb "]
+      end
+    end
+
+  end
+end